tests: move lcr unit test into kam namespace
